+++ controller/index.js
@@ -1,5 +1,7 @@
 var crypto = require('crypto')
 var provider = require('koop-provider')
+var request = require('request')
+var pkg = require('../package.json')
 
 /**
  * creates new gist controller
@@ -7,7 +9,7 @@ var provider = require('koop-provider')
  * @param {object} model - instance of gist model
  */
 var gistController = function (model) {
-  var ctrl = provider.createController()
+  var ctrl = provider.controller()
 
   /**
    * renders index view
@@ -15,9 +17,52 @@ var gistController = function (model) {
    * @param {object} req - incoming request object
    * @param {object} res - outgoing response object
    */
-  function index (req, res) {
+  ctrl.index = function (req, res) {
     res.render(__dirname + '/../views/index', {
       baseUrl: req.baseUrl
+    })
+  }
+
+  /**
+   * renders preview view
+   *
+   * @param {object} req - incoming request object
+   * @param {object} res - outgoing response object
+   */
+  ctrl.preview = function (req, res) {
+    res.render(__dirname + '/../views/demo', {
+      baseUrl: req.baseUrl,
+      id: req.params.id
+    })
+  }
+
+  /**
+   * returns current Github API rate limit
+   *
+   * @param {object} req - incoming request
+   * @param {object} res - outgoing response
+   */
+  ctrl.rate_limit = function (req, res) {
+    var options = {
+      url: 'https://api.github.com/rate_limit',
+      headers: {
+        'User-Agent': 'koop-gist/' + pkg.version
+      }
+    }
+
+    if (model.config.ghtoken) {
+      options.qs = { access_token: model.config.ghtoken }
+    }
+
+    request(options, function (err, response, body) {
+      if (err) {
+        return ctrl.errorResponse({
+          code: response.statusCode,
+          message: err.message
+        }, res)
+      }
+
+      res.jsonp(JSON.parse(body))
     })
   }
 
@@ -28,14 +73,31 @@ var gistController = function (model) {
    * @param {object} req - incoming request object
    * @param {object} res - outgoing response object
    */
-  function find (req, res) {
-    if (!req.params.id) return res.status(404).send('Must specify a user and gist id')
+  ctrl.find = function (req, res) {
+    if (!req.params.id) {
+      return ctrl.errorResponse({
+        code: 422,
+        message: 'Must specify a gist ID'
+      }, res)
+    }
 
     function _send (err, data) {
-      if (err) return res.status(500).json(err)
-      if (!data) return res.status(500).send('There was problem accessing this gist')
-      if (!req.params.format) return res.json(data)
-      if (!model.files.localDir) return res.status(501).send('No local file system configured for exports')
+      if (err) return ctrl.errorResponse({ message: err.message }, res)
+
+      if (!data) {
+        return ctrl.errorResponse({
+          message: 'There was a problem accessing this gist'
+        }, res)
+      }
+
+      if (!req.params.format) return res.jsonp(data)
+
+      if (!model.files.localDir) {
+        return ctrl.errorResponse({
+          code: 501,
+          message: 'No local file system configured for exports'
+        }, res)
+      }
 
       // change geojson to json
       req.params.format = req.params.format.replace('geojson', 'json')
@@ -50,20 +112,25 @@ var gistController = function (model) {
       model.files.exists(path, fileName, function (exists, path) {
         if (exists) {
           if (path.substr(0, 4) === 'http') return res.redirect(path)
-          return res.sendfile(path)
+          return res.sendFile(path)
         }
 
         model.exportToFormat(req.params.format, dir, key, data[0], {}, function (err, file) {
-          if (err) return res.status(500).send(err)
-          res.sendfile(file)
+          if (err) return ctrl.errorResponse({ message: err.message }, res)
+
+          res.sendFile(file)
         })
       })
     }
 
-    model.find(req.params.id, req.query, function (err, data) {
+    model.find({
+      id: req.params.id,
+      query: req.query
+    }, function (err, data) {
       if (!req.params.layer) return _send(err, data)
       if (data[req.params.layer]) return _send(err, data[req.params.layer])
-      _send('Layer not found', null)
+
+      _send(new Error('Layer not found'))
     })
   }
 
@@ -73,37 +140,27 @@ var gistController = function (model) {
    * @param {object} req - incoming request object
    * @param {object} res - outgoing response object
    */
-  function featureservice (req, res) {
-    if (!req.params.id) return res.send('Must specify a gist id', 404)
+  ctrl.featureservice = function (req, res) {
+    if (!req.params.id) {
+      return ctrl.errorResponse({
+        code: 422,
+        message: 'Must specify a gist ID'
+      }, res)
+    }
 
-    var callback = req.query.callback
-    delete req.query.callback
+    model.find({
+      id: req.params.id,
+      query: req.query
+    }, function (err, data) {
+      if (err) return ctrl.errorResponse({ message: err.message }, res)
 
-    model.find(req.params.id, req.query, function (err, data) {
-      ctrl.processFeatureServer(req, res, err, data, callback)
+      // grumble
+      delete req.query.geometry
+      delete req.query.where
+
+      ctrl.processFeatureServer(req, res, data)
     })
   }
-
-  /**
-   * renders preview view
-   *
-   * @param {object} req - incoming request object
-   * @param {object} res - outgoing response object
-   */
-  function preview (req, res) {
-    res.render(__dirname + '/../views/demo', {
-      baseUrl: req.baseUrl,
-      id: req.params.id
-    })
-  }
-
-  /**
-   * assign public functions to controller
-   */
-  ctrl.index = index
-  ctrl.find = find
-  ctrl.featureservice = featureservice
-  ctrl.preview = preview
 
   return ctrl
 }